Measure Unnar’s level of uncertainty
Unnar is sure of many things in his life. It is in fact so rare that Unnar is uncertain, that it is important to measure his level of uncertainty when he is. He says ‘u’ when he is uncertain. Unnar’s level of uncertainty is measured by the duration of his ‘u’. For example, if Unnar says ‘uuuuu’, then his level of uncertainty is 5.

https://open.kattis.com/problems/ovissa